pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airbotics-agent"
version = "2023.08.08"
description = "The Airbotics cloud robotics agent"
readme = "README.md"
authors = [{ name = "Airbotics Inc.", email = "admin@airbotics.io" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]
keywords = ["airbotics", "agent", "robotics", "ros"]
dependencies = [
    "python-daemon>=2.3.0",
    "PyYAML>=6.0",
    "docker>=5.0.3",
    "paho-mqtt>=1.6.1",
    "requests>=2.27.1",
    "backoff>=2.0.1",
    "jsonschema>=4.6.0",
    "pyfiglet>=0.8.post1",
    "numpy>=1.25.0",
    "toml>=0.10.2",
    "distro>=1.7.0",
    "psutil>=5.9.5",
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/Airbotics"

[project.scripts]
airbotics = "agent.__main__:main"
